remove paymentCardTypeCode from mapToZeroDollarRefundRequestData
