+++ frontend/src/features/permits/pages/Amend/components/AmendPermitFinish.tsx
@@ -96,7 +96,7 @@ export const AmendPermitFinish = () => {
 
     if (isZeroAmount(amountToRefund)) {
       startTransactionMutation.mutate(
-        mapToZeroDollarRefundRequestData(refundData, permitId),
+        mapToZeroDollarRefundRequestData(permitId),
       );
     } else {
       refundPermitMutation.mutate({
